Added preliminary support for Hardchord Music block format
Added support for the first revision of the Harchord Music block
format, which is essentially a byte array that calls all the
functions for playing music.

The format is promising, reducing one of our 60-second song
demos from 24,870 bytes to 10,246. However, it has a lot of
overhead since it addresses many functions that would usually
be compiled out. We recommend you only use it in projects that
store a *lot* of pre-written music.

This release also converts the private _shiftOut back to a
static void, saving more space.