Removed unnecessary object specifiers
This update removes all the unnecessary object specifiers (why
the hell did I have so many "this->"?).

More importantly, it changes the behavior of beat() to take a
uint8_t as the second parameter instead of a float. This helps
to shrink compiled code size a bit. We now pass a numerator to
be resolved over a denominator of 8 in the function.